// src/types/bindings.ts
export const events = {
	downloadAnimeError: makeEvent<DownloadAnimeError>("download-anime-error"),
	downloadAnimeInfo: makeEvent<DownloadAnimeInfo>("download-anime-info"),
	downloadCancelled: makeEvent<DownloadCancelled>("download-cancelled"),
	downloadDone: makeEvent<DownloadDone>("download-done"),
	downloadError: makeEvent<DownloadError>("download-error"),
	downloadInfo: makeEvent<DownloadInfo>("download-info"),
	downloadProgress: makeEvent<DownloadProgress>("download-progress"),
};
// ...
export type DownloadAnimeError = {
	anime_id: string,
	episode: EpisodeInformation,
	resolution: Resolution,
	message: string,
};

export type DownloadAnimeInfo = {
	anime_id: string,
	episode: EpisodeInformation,
	completed: boolean,
	resolution: Resolution,
};

export type DownloadCancelled = {
	id: string,
	resolution: Resolution,
};

export type DownloadDone = {
	id: string,
	resolution: Resolution,
	path: string,
};

export type DownloadError = {
	id: string,
	resolution: Resolution,
	message: string,
};

export type DownloadInfo = {
	id: string,
	resolution: Resolution,
	total_size: number | null,
};

export type DownloadProgress = {
	id: string,
	resolution: Resolution,
	downloaded: number | null,
	total: number | null,
};

export type EpisodeInformation = {
	name: string,
	date: string,
	url: string,
};

export type EpisodeStreamingInformation = {
	episode_information: EpisodeInformation,
	anime_name: string,
	anime_folder: string,
	episodes: EpisodeInformation[],
	default_mirror: string,
	mirrors: ResolutionMirror[],
};
// ...
export type Mirror = {
	name: string,
	data: MirrorData,
};

export type MirrorData = {
	id: number,
	i: number,
	q: string,
};

export type MirrorLink = {
	url: string,
	nonce: string,
};

export type Platform = "Pdrain" | "Mega" | "DesuDrive";

export type RenameMapping = {
	original_path: string,
	new_name: string,
};

export type Resolution = "P360" | "P480" | "P720";

export type ResolutionMirror = {
	resolution: string,
	mirrors: Mirror[],
};